requirements.txt
DrissionPage>=4.1.0
openai>=1.30.0
pyyaml>=6.0
python-dotenv>=1.0.0
